Refactoring and tests

+++ src/main/java/edu/hw1/Task5.java
@@ -3,7 +3,7 @@ package edu.hw1;
 import edu.hw1.utils.DigitsHelper;
 
 public class Task5 {
-    private static final int BASE_10 = 10;
+    private static final short BASE_10 = 10;
 
     private Task5() {
     }
@@ -11,7 +11,7 @@ public class Task5 {
     public static boolean isPalindromeDescendant(long number) {
         long positive = Math.abs(number);
         int digitsCount = Task2.countDigits(positive);
-        int[] digits = DigitsHelper.toDigitsArray(positive, digitsCount, BASE_10);
+        short[] digits = DigitsHelper.toDigitsArray(positive, digitsCount, BASE_10);
 
         do {
             if (isPalindrome(digits, digitsCount)) {
@@ -24,7 +24,7 @@ public class Task5 {
         return false;
     }
 
-    private static int transformToItsChildAndReturnNewCount(int[] digits, int digitsCount) {
+    private static int transformToItsChildAndReturnNewCount(short[] digits, int digitsCount) {
         int newCount = 0;
         int i = 1;
 
@@ -34,19 +34,19 @@ public class Task5 {
         }
 
         for (; i < digitsCount; i += 2) {
-            int sum = digits[i] + digits[i - 1];
+            short sum = (short) (digits[i] + digits[i - 1]);
             if (sum < BASE_10) {
                 digits[newCount++] = sum;
             } else {
-                digits[newCount++] = sum / BASE_10;
-                digits[newCount++] = sum % BASE_10;
+                digits[newCount++] = (short) (sum / BASE_10);
+                digits[newCount++] = (short) (sum % BASE_10);
             }
         }
 
         return newCount;
     }
 
-    private static boolean isPalindrome(int[] array, int count) {
+    private static boolean isPalindrome(short[] array, int count) {
         for (int i = 0; i < count / 2; i++) {
             int oppositeIndex = count - 1 - i;
             if (array[i] != array[oppositeIndex]) {
